requirements.txt
matplotlib==3.4.2
seaborn==0.11.1
pandas==1.2.4
altair==4.1.0
numpy==1.20.3
streamlit==0.82.0
Pillow==8.2.0
scikit_learn==0.24.2
